mcp-crash-course-project-sse/pyproject.toml
[project]
name = "mcp-crash-course-project-sse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.12",
    "langgraph>=0.3.25",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
]
